Create combat_animation, hit_animation, moving_animation, preloaded_asset_sprite, sprite and sprite_sequence directories.
