+++ src/classes/animation/combat_animation/brigand_combat_animation.js
@@ -1,5 +1,5 @@
-import CombatSprite from '../../../../animations/spriteSequence/combatSprite';
-import Sprite from '../../../../animations/sprite';
+import CombatSprite from '@/animation/sprite_sequence/combat_sprite';
+import Sprite from '@/animation/sprite/sprite';
 
 function BrigandCombatAnimation() {
   let sprites0 = new Sprite(70, 39, 70, 39, 'src/assets/combat_spritesheets/brigand/brigandCombatSprite0.png', 6, 4);
