+++ src/classes/animation/combat_animation/lyn_crit_combat_animation.js
@@ -1,5 +1,5 @@
-import CombatSprite from '../../../../animations/spriteSequence/combatSprite';
-import Sprite from '../../../../animations/sprite';
+import CombatSprite from '@/animation/sprite_sequence/combat_sprite';
+import Sprite from '@/animation/sprite/sprite';
 
 function LynCritCombatAnimation() {
   let sprite0 = new Sprite(90, 50, 90, 50, 'src/assets/combat_spritesheets/lyn/lynCombatSprite0.png', 4, 5);
