+++ src/classes/animation/combat_animation/roy_receive_hit_sprite.js
@@ -1,5 +1,5 @@
-import SpriteSequence from '../../../../animations/spriteSequence/spriteSequence';
-import Sprite from '../../../../animations/sprite';
+import SpriteSequence from '@/animation/sprite_sequence/sprite_sequence';
+import Sprite from '@/animation/sprite/sprite';
 
 function RoyReceiveHitAnimation() {
   const sprite0 = new Sprite(70, 43, 70, 43, 'src/assets/combat_spritesheets/roy/royReceiveHitSprite.png', 4, 1);
