+++ src/classes/animation/hit_animation/normal_damage_animation.js
@@ -1,15 +1,15 @@
-import CoordinateSprite from '../coordinateSprite';
+import CoordinateSprite from '@/animation/sprite/coordinate_sprite';
 
-function NormalCriticalDamageAnimation(defenderCoordinates) {
+function NormalDamageAnimation(defenderCoordinates) {
   CoordinateSprite.call(
     this,
-    240,//width
-    143,//height
+    238,//width
+    158,//height
     //238,//renderWidth
-    240*1.4,//renderWidth
+    238*1.8,//renderWidth
 
-    //158,//renderHeight
-    143*1.3,//renderHeight
+    // 158,//renderHeight
+    158*1.8,//renderHeight
 
     'src/assets/hit_effect_spritesheets/FE7HitSpriteSheetsTransTest.png',//image
     2,//ticksPerFrame
@@ -17,29 +17,27 @@ function NormalCriticalDamageAnimation(defenderCoordinates) {
     [
       // [10, 21],
       // [253, 1033],
-      // //[10, 1197],
-      // //[253, 1033],
+      // [10, 1197],
+      // [253, 1033],
       // [10,21],
 
-      [253, 526], [496, 526],
-      [10, 690], [253, 690], [496, 690],
-      [10, 853], [253, 853], [496, 853]
-
-    ]//coordinatesList
+     [253, 21], [496, 21],
+    [10, 184], [253, 184], [496, 184],
+    [10, 347], [253, 347], [496, 347]]//coordinatesList
   );
 
   this.defenderCoordinates = defenderCoordinates;
 }
 
-NormalCriticalDamageAnimation.prototype = Object.create(CoordinateSprite.prototype);
-NormalCriticalDamageAnimation.prototype.constructor = NormalCriticalDamageAnimation;
+NormalDamageAnimation.prototype = Object.create(CoordinateSprite.prototype);
+NormalDamageAnimation.prototype.constructor = NormalDamageAnimation;
 
-NormalCriticalDamageAnimation.prototype.render = function(sF) {
+NormalDamageAnimation.prototype.render = function(sF) {
   let scale = sF / 18;
   // let cx = (this.defenderCoordinates[0] * sF) + (((scale * 18) - (scale * this.renderWidth)) / 2);
   // let cy = (this.defenderCoordinates[1] * sF) + ((scale * 18) - (scale * this.renderHeight));
-  let cx = (this.defenderCoordinates[0] * sF) + (((scale * 18) - (scale * this.renderWidth)) / 2) + (1.5*sF);
-  let cy = (this.defenderCoordinates[1] * sF) + ((scale * 18) - (scale * this.renderHeight)) + (2*sF);
+  let cx = (this.defenderCoordinates[0] * sF) + (((scale * 18) - (scale * this.renderWidth)) / 2) + (1.2*sF);
+  let cy = (this.defenderCoordinates[1] * sF) + ((scale * 18) - (scale * this.renderHeight)) + (sF*2.5);
 
   let cWidth = scale * this.renderWidth;
   let cHeight = scale *  this.renderHeight;
@@ -57,4 +55,4 @@ NormalCriticalDamageAnimation.prototype.render = function(sF) {
   this.update();
 }
 
-export default NormalCriticalDamageAnimation;
+export default NormalDamageAnimation;
