+++ src/classes/animation/moving_animation/moving_animation.js
