+++ src/classes/animation/preloaded_asset_sprite/battle_asset_image.js
