+++ src/classes/animation/sprite/battle_platform_sprite.js
@@ -1,4 +1,4 @@
-import CoordinateSprite from './coordinateSprite';
+import CoordinateSprite from '@/animation/sprite/coordinate_sprite';
 
 function BattlePlatformSprite() {
   CoordinateSprite.call(
