+++ src/classes/attack/pu_attack.js
@@ -1,7 +1,7 @@
 import Attack from './attack';
-import NormalDamageAnimation from '../../../animations/hitAnimations/normalDamageAnimation';
-import NormalCriticalDamageAnimation from '../../../animations/hitAnimations/normalCriticalDamageAnimation';
-import RoyCriticalHitAnimation from '../../../animations/hitAnimations/royCriticalHitAnimation';
+import NormalDamageAnimation from '@/animation/hit_animation/normal_damage_animation';
+import NormalCriticalDamageAnimation from '@/animation/hit_animation/roy_critical_hit_animation';
+import RoyCriticalHitAnimation from '@/animation/hit_animation/roy_critical_hit_animation';
 import Roy from '@/unit/roy';
