+++ src/classes/player/enemy_player.js
@@ -2,9 +2,9 @@ import GeneralPlayer from  '@/player/general_player';
 import PlayerUnit from '@/unit/player_unit';
 import EnemyUnit from '@/unit/enemy_unit';
 import { equivalentPositions } from '~/util';
-import MovingAnimation from '../../../animations/movingAnimation';
+import MovingAnimation from '@/animation/moving_animation/moving_animation';
 import Combat from '@/combat/combat';
-import CombatAnimation from '../../../animations/combatAnimation';
+import CombatAnimation from '@/animation/combat_animation/combat_animation';
 
 function EnemyPlayer(board, display, phaseStage, frameSource) {
   this.board = board;
